+++ Test/VideoTranscriptTests.swift
@@ -1,5 +1,5 @@
 //
-//  OEXVideoTranscriptTests.swift
+//  VideoTranscriptTests.swift
 //  edX
 //
 //  Created by Danial Zahid on 1/11/17.
@@ -9,36 +9,38 @@
 import XCTest
 @testable import edX
 
-class OEXVideoTranscriptTests: XCTestCase {
+class VideoTranscriptTests: XCTestCase {
     
     func testTranscriptLoaded() {
-        let transcriptView = OEXVideoTranscript()
+        let environment = TestRouterEnvironment()
+        let transcriptView = VideoTranscript(environment: environment)
         
         XCTAssertEqual(transcriptView.transcriptTableView.numberOfRowsInSection(0), 0)
         XCTAssertTrue(transcriptView.transcriptTableView.hidden)
         
-        transcriptView.updateTranscript(OEXVideoTranscriptDataFactory.transcriptArray())
+        transcriptView.updateTranscript(VideoTranscriptDataFactory.transcriptArray())
         
         XCTAssertEqual(transcriptView.transcriptTableView.numberOfRowsInSection(0), 5)
         XCTAssertFalse(transcriptView.transcriptTableView.hidden)
     }
     
     func testTransctipSeek() {
-        let transcriptView = OEXVideoTranscript()
-        transcriptView.updateTranscript(OEXVideoTranscriptDataFactory.transcriptArray())
+        let environment = TestRouterEnvironment()
+        let transcriptView = VideoTranscript(environment: environment)
+        transcriptView.updateTranscript(VideoTranscriptDataFactory.transcriptArray())
         
         transcriptView.highlightSubtitleForTime(1.75)
-        XCTAssertEqual(transcriptView.selectedIndex, 1)
+        XCTAssertEqual(transcriptView.highlightedIndex, 1)
         
         transcriptView.highlightSubtitleForTime(3.45)
-        XCTAssertEqual(transcriptView.selectedIndex, 2)
+        XCTAssertEqual(transcriptView.highlightedIndex, 2)
         
         transcriptView.highlightSubtitleForTime(3.47)
-        XCTAssertEqual(transcriptView.selectedIndex, 2)
+        XCTAssertEqual(transcriptView.highlightedIndex, 2)
     }
 }
 
-class OEXVideoTranscriptDataFactory{
+class VideoTranscriptDataFactory{
     
     static func transcriptArray() -> [AnyObject] {
         var transcript = [[String: AnyObject]]()
